+++ .gitignore
@@ -12,6 +12,7 @@
 *.out
 *.tfstate
 *.hcl
+*.backup
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
