+++ docs/index.md
@@ -1,0 +1,21 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "plugin Provider"
+subcategory: ""
+description: |-
+  
+---
+
+# plugin Provider
+
+
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `seed` (String) seed for generating keys of provider service API.
+- `url` (String) server url of provider service API.
