+++ docs/resources/pkid_key_query.md
@@ -1,0 +1,42 @@
+---
+# generated by https://github.com/hashicorp/terraform-plugin-docs
+page_title: "plugin_pkid_key_query Resource - terraform-provider-plugin"
+subcategory: ""
+description: |-
+  
+---
+
+# plugin_pkid_key_query (Resource)
+
+
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Required
+
+- `encrypt` (Boolean)
+- `key` (String)
+- `project` (String)
+- `value` (String)
+
+### Optional
+
+- `timeouts` (Block, Optional) (see [below for nested schema](#nestedblock--timeouts))
+
+### Read-Only
+
+- `id` (String) The ID of this resource.
+
+<a id="nestedblock--timeouts"></a>
+### Nested Schema for `timeouts`
+
+Optional:
+
+- `create` (String)
+- `delete` (String)
+- `update` (String)
+
+
